refactored debug component
