+++ components/debug/index.js
@@ -1,0 +1,3 @@
+import { APIComponent } from './APIComponent';
+
+export { APIComponent };
